--- src/app/(base)/auth/page.tsx
@@ -1,5 +1,0 @@
-import AuthComp from "@/components/AuthComp";
-
-export default function page() {
-  return <AuthComp/>;
-}
