+++ src/app/[id]/1stgame/ral0P/page.tsx
@@ -1,7 +1,7 @@
 import FirstGameForm from "@/components/FirstGameForm";
-import findOrder from "@/lib/findOrder";
 import { validateSessionID } from "@/lib/validateSessionId";
 import { notFound } from "next/navigation";
+import findOrder from "@/lib/findOrder";
 
 export default async function page({ params }: { params: { id: string } }) {
   //read sessionId in url and validate it
@@ -14,13 +14,11 @@ export default async function page({ params }: { params: { id: string } }) {
   const assessCond = await findOrder(sessionID);
 
   return (
-    <div>
-      <FirstGameForm
-        sessionId={sessionID}
-        condition="angry"
-        money={"250"}
-        assess_cond={assessCond}
-      />
-    </div>
+    <FirstGameForm
+      sessionId={sessionID}
+      condition="neutral"
+      money={"500"}
+      assess_cond={assessCond}
+    />
   );
 }
