+++ src/app/[id]/2ndgame/dYezj/page.tsx
