+++ src/app/[id]/2ndgame/nbXj6/page.tsx
