+++ src/app/[id]/2ndgame/sO852/page.tsx
