+++ src/app/[id]/debriefing/page.tsx
@@ -1,7 +1,6 @@
-import FirstGameForm from "@/components/FirstGameForm";
+import AgreementForm from "@/components/AgreementForm";
 import { validateSessionID } from "@/lib/validateSessionId";
 import { notFound } from "next/navigation";
-import findOrder from "@/lib/findOrder";
 
 export default async function page({ params }: { params: { id: string } }) {
   //read sessionId in url and validate it
@@ -11,16 +10,5 @@ export default async function page({ params }: { params: { id: string } }) {
     notFound();
   }
 
-  const assessCond = await findOrder(sessionID);
-
-  return (
-    <div>
-      <FirstGameForm
-        sessionId={sessionID}
-        condition="neutral"
-        money={"500"}
-        assess_cond={assessCond}
-      />
-    </div>
-  );
+  return <AgreementForm id={sessionID} />;
 }
