+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
-import "../../../../public/styles/globals.css";
+import "../../public/styles/globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
+import Link from "next/link";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -18,24 +19,12 @@ export default function RootLayout({
     id: string;
   };
 }) {
+  console.log(params.id);
   return (
     <html lang="en">
       <body className={inter.className}>
         <div className="flex flex-col min-h-screen">
-          <header className="border-b-2 border-black font-bold tracking-tight w-full">
-            <div className="mx-3 sm:container sm:mx-auto">
-              <nav className="flex justify-between items-center">
-                <div>
-                  <h1 className="text-lg sm:text-4xl my-3">
-                    応用経済分析研究室
-                  </h1>
-                </div>
-
-              </nav>
-            </div>
-          </header>
-
-          <div className="mx-3 sm:container sm:mx-auto flex-1">{children}</div>
+          {children}
 
           <footer className="mt-12 border-t-2 border-black w-full">
             <div className="mx-3 sm:container sm:mx-auto my-3">
